ログイン機能をContextを使うように修正
